Move elasticsearch collector directory so it won't overlap with the elasticsearch module.
NOTE: Class name stays the same.
