+++ src/collectors/elasticsearch_collector/elasticsearch_collector.py
