+++ src/collectors/elasticsearch_collector/test/fixtures/cluster_stats
