+++ src/collectors/elasticsearch_collector/test/fixtures/indices_stats
