+++ src/collectors/elasticsearch_collector/test/fixtures/logstash_indices_stats
