+++ src/collectors/elasticsearch_collector/test/fixtures/stats
