+++ src/collectors/elasticsearch_collector/test/fixtures/stats0.90
